refactoring: moved all kjTree to C++ object functions to own modules in kjTree lib
