+++ test/functionalTest/cases/0000_ngsild/ngsild_big_context.test
@@ -61,7 +61,7 @@ payload='{
     "object": "urn:ngsi-ld:PointOfInterest:RZ:MainSquare"
   },
   "@context": [
-    "https://schema.lab.fiware.org/ld/jsonldcontext.jsonld",
+    "https://schema.lab.fiware.org/ld/context.jsonld",
     "https://uri.etsi.org/ngsi-ld/v1/ngsi-ld-core-context.jsonld"
   ]
 }'
