+++ test/functionalTest/cases/0000_ngsild/ngsild_issue_0065.test
@@ -47,7 +47,7 @@ echo "==========================================================================
 payload='{
    "@context": [
             "https://uri.etsi.org/ngsi-ld/v1/ngsi-ld-core-context.jsonld",
-            "https://schema.lab.fiware.org/ld/jsonldcontext.jsonld"
+            "https://schema.lab.fiware.org/ld/context.jsonld"
         ],
     "id": "urn:ngsi-ld:Room:Room1",
     "type": "Room",
